requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
black==22.12.0
bleach==6.0.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.4
coverage==7.3.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
display==1.0.0
distlib==0.3.6
distro==1.8.0
Django==4.2
djangorestframework==3.14.0
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.0
fqdn==1.5.1
frozenlist==1.4.0
geosky==1.0.9
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.8.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
mypy-extensions==0.4.3
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.24.3
openai==1.2.3
overrides==7.4.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pickleshare==0.7.5
platformdirs==3.5.1
pluggy==1.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.5.2
Pygments==2.14.0
pytest==7.2.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.3
terminado==0.17.1
tiktoken==0.5.1
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.8.0
typing_extensions==4.8.0
tzdata==2023.3
unicodecsv==0.14.1
Unidecode==1.3.6
uri-template==1.3.0
urllib3==1.26.14
virtualenv==20.23.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Wikipedia-API==0.5.8
yarl==1.9.2
yelpapi==2.5.0
